update user setting
