+++ app/src/main/AndroidManifest.xml
@@ -16,12 +16,6 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".buotg.view.user_setting.UserInfoEditActivity"
-            android:exported="false" />
-        <activity
-            android:name=".buotg.view.user_setting.StudentLinkActivity"
-            android:exported="false" />
-        <activity
             android:name=".buotg.view.user_setting.SettingActivity"
             android:exported="false" />
         <activity
