--- app/src/main/java/com/cs501/cs501app/buotg/view/user_setting/StudentLinkActivity.kt
@@ -1,78 +1,0 @@
-package com.cs501.cs501app.buotg.view.user_setting
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.cs501.cs501app.buotg.view.user_setup.StuLinkImport
-import com.cs501.cs501app.utils.GenericTopAppBar
-
-class StudentLinkActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            MaterialTheme {
-                RenderScaffold()
-            }
-        }
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    fun RenderScaffold() {
-        val steps2 = listOf( "Import from studentlink")
-        val (step, setStep) = remember { mutableStateOf(0) }
-        val title = "Student Link"
-        // callback for when a step is done
-        val stepDone: () -> Unit = {
-            Log.d("SetupActivity", "stepDone: step = $step, steps2.size = ${steps2.size}")
-            if (step < steps2.size - 1) {
-                setStep(step + 1)
-            }
-        }
-        Scaffold(
-            topBar = {
-                GenericTopAppBar(title = title)
-            }
-        ) { innerPadding ->
-            Column(
-                modifier = Modifier
-                    .padding(innerPadding)
-                    .padding(16.dp)
-                    .fillMaxWidth()
-                    .fillMaxHeight()
-                    .verticalScroll(rememberScrollState()),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Top,
-            ) {
-                // render the current step
-                when (step) {
-                    0 -> StuLinkImport(stepDone)
-                }
-                // spacer to keep things to the bottom
-                // https://stackoverflow.com/questions/70904979/how-align-to-bottom-a-row-in-jetpack-compose
-                Spacer(modifier = Modifier.weight(1f))
-                // row of buttons, sticks to the bottom
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.Bottom,
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-
-                }
-            }
-        }
-    }
-}
