--- app/src/main/java/com/cs501/cs501app/buotg/view/user_setting/UserInfoEditActivity.kt
@@ -1,154 +1,0 @@
-package com.cs501.cs501app.buotg.view.user_setting
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.cs501.cs501app.buotg.database.SyncRepo
-import com.cs501.cs501app.buotg.database.entities.User
-import com.cs501.cs501app.buotg.database.entities.UserType
-import com.cs501.cs501app.buotg.database.repositories.AppRepository
-import com.cs501.cs501app.buotg.view.user_setup.SetupActivity
-import com.cs501.cs501app.utils.GenericTopAppBar
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import java.util.*
-
-class UserInfoEditActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            MaterialTheme {
-                UserInfoEditScreen()
-            }
-        }
-    }
-
-    @Composable
-    @OptIn(ExperimentalMaterial3Api::class)
-    fun UserInfoEditScreen() {
-        val userRepo = AppRepository.get().userRepo()
-        val coroutineScope = rememberCoroutineScope()
-//        val currentUser = remember { mutableStateOf<User?>(null) }
-        val currentUser = User(
-            UUID.randomUUID(),
-            "example@example.com",
-            "John Doe",
-            Date(),
-            UserType.student
-        )
-        // Fetch the current user info
-//        LaunchedEffect(key1 = Unit) {
-//            currentUser.value = withContext(Dispatchers.IO) {
-//                userRepo.getCurrentUser()
-//            }
-//        }
-
-        Scaffold(
-            topBar = {
-                GenericTopAppBar()
-            },
-            content = { innerPadding ->
-                Column(
-                    modifier = Modifier
-                        .padding(innerPadding)
-                        .padding(16.dp)
-                        .fillMaxWidth()
-                        .fillMaxHeight()
-                        .verticalScroll(rememberScrollState()),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center,
-                ) {
-
-//                    currentUser.value?.let { user ->
-//                        // Show the user's info in a form
-//                        UserInfoForm(user = user) { updatedUser ->
-//                            // Update the user info in the database
-//                            coroutineScope.launch {
-//                                withContext(Dispatchers.IO) {
-//                                    userRepo.updateUser(updatedUser)
-//                                }
-//                                Toast.makeText(this@UserInfoEditActivity, "User info updated", Toast.LENGTH_SHORT).show()
-//                                finish()
-//                            }
-//                        }
-//                    }
-
-                    currentUser?.let { user ->
-                        // Show the user's info in a form
-                        UserInfoForm(user = user) { updatedUser ->
-                            // Update the user info in the database
-                            coroutineScope.launch {
-                                withContext(Dispatchers.IO) {
-                                    userRepo.updateUser(updatedUser)
-                                }
-                                Toast.makeText(this@UserInfoEditActivity, "User info updated", Toast.LENGTH_SHORT).show()
-                                finish()
-                            }
-                        }
-                    }
-                }
-            }
-        )
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    fun UserInfoForm(user: User, onUserInfoUpdated: (User) -> Unit) {
-        val full_name = remember { mutableStateOf(user.full_name) }
-        val email = remember { mutableStateOf(user.email) }
-
-        Column(
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            OutlinedTextField(
-                value = full_name.value,
-                onValueChange = { full_name.value = it },
-                label = { Text("Name") },
-                singleLine = true
-            )
-
-            OutlinedTextField(
-                value = email.value,
-                onValueChange = { email.value = it },
-                label = { Text("Email") },
-                singleLine = true
-            )
-
-
-            Button(
-                onClick = {
-                    // Create a new user object with the updated info
-                    val updatedUser = User(
-                        created_at = user.created_at,
-                        full_name = full_name.value,
-                        email = email.value,
-                        user_type = user.user_type,
-                        user_id = user.user_id
-                    )
-
-                    // Call the onUserInfoUpdated callback to notify the caller of the updated user info
-                    onUserInfoUpdated(updatedUser)
-                }
-            ) {
-                Text("Save")
-            }
-        }
-    }
-}
-
